--- a/torch/csrc/utils/python_arg_parser.h
+++ b/torch/csrc/utils/python_arg_parser.h
@@ -179,7 +179,7 @@ struct PYBIND11_EXPORT PythonArgParser {
   bool traceable;
 };
 
-struct PYBIND11_EXPORT FunctionSignature {
+struct FunctionSignature {
   explicit FunctionSignature(const std::string& fmt, int index);
 
   // NOLINTNEXTLINE(cppcoreguidelines-avoid-c-arrays,modernize-avoid-c-arrays)
@@ -188,13 +188,13 @@ struct PYBIND11_EXPORT FunctionSignature {
       PyObject* args,
       PyObject* kwargs,
       PyObject* dst[],
+      std::vector<py::handle>& overloaded_args,
       bool raise_exception);
 
   std::string toString() const;
 
   std::string name;
   std::vector<FunctionParameter> params;
-  std::vector<py::handle> overloaded_args;
   size_t min_args;
   size_t max_args;
   size_t max_pos_args;
@@ -203,20 +203,23 @@ struct PYBIND11_EXPORT FunctionSignature {
   bool deprecated;
 };
 
-struct PythonArgs {
+struct PYBIND11_EXPORT PythonArgs {
   PythonArgs(
       bool traceable,
       const FunctionSignature& signature,
-      PyObject** args)
+      PyObject** args,
+      std::vector<py::handle> overloaded_args)
       : idx(signature.index),
         traceable(traceable),
         signature(signature),
-        args(args) {}
+        args(args),
+        overloaded_args(std::move(overloaded_args)) {}
 
   int idx;
   bool traceable;
   const FunctionSignature& signature;
   PyObject** args;
+  std::vector<py::handle> overloaded_args;
 
   inline bool has_torch_function();
   inline std::string get_func_name();
@@ -364,7 +367,7 @@ inline PythonArgs PythonArgParser::parse(PyObject* self, ParsedArgs<0>& dst) {
 }
 
 inline bool PythonArgs::has_torch_function() {
-  return !this->signature.overloaded_args.empty() ||
+  return !overloaded_args.empty() ||
       at::impl::torch_function_mode_enabled();
 }