@colesbury 2023-08-15
FYI, I've changed `std::vector<py::handle>` to `std::vector<PyObject*>` in `struct PythonArgs`. The use of `py::handle` in the returned struct caused symbol visibility and linker problems that I wasn't able to resolve. (See e.g., https://github.com/pytorch/pytorch/actions/runs/5860835115/job/15892629532).